put MqttClient in a module
